Fixed blockchain bugs updated PoW system to properly compare hashes

+++ internal/transactions/transactions.go
@@ -6,5 +6,5 @@ type Transaction struct {
 	ID        uuid.UUID
 	Sender    string
 	Recipient string
-	Amount    float64
+	Amount    float32
 }
